# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product Service API",
        "description": "API документация для сервиса товаров",
        "contact": {
            "email": "yuriyseraz@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Поиск товаров",
                "description": "Возвращает список товаров с пагинацией, фильтрацией и сортировкой.",
                "operationId": "getProductsList",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Поиск по названию (LIKE)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "ID категории",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "price_from",
                        "in": "query",
                        "description": "Цена от",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "price_to",
                        "in": "query",
                        "description": "Цена до",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "in_stock",
                        "in": "query",
                        "description": "Только в наличии (1 - да, 0 - нет)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "rating_from",
                        "in": "query",
                        "description": "Рейтинг от (0-5)",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Сортировка",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "newest",
                            "enum": [
                                "price_asc",
                                "price_desc",
                                "rating_desc",
                                "newest"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Кол-во товаров на страницу",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ProductResource"
                                            }
                                        },
                                        "links": {
                                            "properties": {
                                                "first": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "last": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "next": {
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "from": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "path": {
                                                    "type": "string"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "to": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The category id field is invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProductResource": {
                "title": "Product",
                "description": "Схема товара",
                "required": [
                    "id",
                    "name",
                    "price",
                    "category"
                ],
                "properties": {
                    "id": {
                        "description": "ID товара",
                        "type": "integer",
                        "example": 101
                    },
                    "name": {
                        "description": "Название товара",
                        "type": "string",
                        "example": "Iphone 15"
                    },
                    "price": {
                        "description": "Цена",
                        "type": "number",
                        "format": "float",
                        "example": 999.99
                    },
                    "in_stock": {
                        "description": "Наличие на складе",
                        "type": "boolean",
                        "example": true
                    },
                    "category": {
                        "description": "Категория товара",
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 5
                            },
                            "name": {
                                "type": "string",
                                "example": "Electronics"
                            }
                        },
                        "type": "object"
                    },
                    "can_buy": {
                        "description": "Доступен ли для покупки (бизнес-логика)",
                        "type": "boolean",
                        "example": true
                    },
                    "rating": {
                        "description": "Рейтинг",
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "created_at": {
                        "description": "Дата создания",
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-20"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}